+++ packages/veui-theme-dls/components/Pagination.js
@@ -1,7 +1,7 @@
 import 'veui-theme-dls-icons/chevron-left'
 import 'veui-theme-dls-icons/chevron-right'
-import 'veui-theme-one-icons/angle-double-left'
-import 'veui-theme-one-icons/angle-double-right'
+import 'veui-theme-dls-icons/chevron-double-left'
+import 'veui-theme-dls-icons/chevron-double-right'
 import 'veui-theme-dls-icons/ellipsis'
 import config from 'veui/managers/config'
 
@@ -10,8 +10,8 @@ config.defaults(
     icons: {
       prev: 'chevron-left',
       next: 'chevron-right',
-      backward: 'angle-double-left',
-      forward: 'angle-double-right',
+      backward: 'chevron-double-left',
+      forward: 'chevron-double-right',
       more: 'ellipsis'
     },
     ui: {